backend/requirements.txt
Django>=2.0,<3.0
psycopg2-binary>=2.8
django-crispy-forms
djangorestframework
djangorestframework-jwt
django-cors-headers
coverage